+++ tests/test_simulation.py
@@ -1,10 +1,6 @@
 import pathlib
 
-from pyechelle import simulator, benchmark
-
-
-# disable jit, because it is not
-# os.environ['NUMBA_DISABLE_JIT'] = '4'
+from pyechelle import simulator
 
 
 def test_simulation(capsys, benchmark):
@@ -33,6 +29,10 @@ def test_simulation_multicore(capsys, benchmark):
 
 
 def test_benchmark():
-    benchmark.run_benchmark_cpu([1], [0.01])
+    # import os
+    # os.environ['NUMBA_DISABLE_JIT'] = '0'
+    from pyechelle import benchmark
+
+    benchmark.run_benchmark_cpu([1], [0.00001])
     # benchmark.run_benchmark_cuda([0.01])
     pathlib.Path.cwd().joinpath('test.fits').unlink(missing_ok=True)