+++ tests/test_simulation.py
@@ -1,5 +1,7 @@
 import pathlib
 
+from astropy.io import fits
+
 from pyechelle import simulator
 
 
@@ -11,6 +13,8 @@ def test_simulation(capsys, benchmark):
     captured = capsys.readouterr()
     result = captured.out
     assert "Simulation took" in result
+    # check that metadata has been written to fits file
+    assert 'pyechelle.' in ''.join([h for h in fits.getheader('test.fits')])
     # cleanup files
     pathlib.Path.cwd().joinpath('test.fits').unlink(missing_ok=True)