+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.7] - unreleased
+
+### Added
+- Simulation metadata is now added to FITS header (fits files now contain information about simulation parameters)
+
+### Fixed
+ - fixed #9 (issue with CUDA random seed)
+
+
 ## [0.3.6] - 2023-08-07
 
 ### Added
@@ -27,7 +36,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
-- fixed #8
+- fixed #8 (atmospheric absorption was not applied)
 
 ## [0.3.4] - 2022-08-28